Graphing now kind of works, need to add headers and improve layout.

+++ Interface/DataPrepper.cs
@@ -7,7 +7,7 @@ namespace Interface;
 
 public static class DataPrepper {
     /// <summary>
-    /// Converts an .xls file to a .xlsx
+    ///     Converts an .xls file to a .xlsx
     /// </summary>
     /// <param name="path">Path to .xls</param>
     public static void Convert(string path) {
@@ -28,15 +28,13 @@ public static class DataPrepper {
     }
 
     /// <summary>
-    /// Preps samples from the tensile machine.
+    ///     Preps samples from the tensile machine.
     /// </summary>
     /// <param name="inputDir">Directory where the .xls files are stored</param>
     /// <param name="recursive">True to also search subdirectories</param>
     /// <param name="progressUpdate">Delegate which is called every time a sheet is converted</param>
     public static void PrepTensile(string inputDir, bool recursive, Action<double> progressUpdate) {
-        if (Directory.Exists("temp")) {
-            Directory.Delete("temp", true);
-        }
+        if (Directory.Exists("temp")) Directory.Delete("temp", true);
         Directory.CreateDirectory("temp");
         ExcelPackage.LicenseContext = LicenseContext.NonCommercial;
         Encoding.RegisterProvider(CodePagesEncodingProvider.Instance);
